Fix the timestamp overflow bug
This patch fix an Android defeat. 32 bits timestamp is not big enough,
change to 64 bits to avoid overflow.

Author: Wei Feng <wei.feng@intel.com>
Author-tracking-BZ: 28089, 50102

+++ media/libstagefright/include/SampleTable.h
@@ -65,7 +65,7 @@ public:
             uint32_t sampleIndex,
             off64_t *offset,
             size_t *size,
-            uint32_t *compositionTime,
+            int64_t *compositionTime,
             bool *isSyncSample = NULL);
 
     enum {
